# chore(zql): explore range query perf