Run:

> `npm run test | glow`
![CleanShot 2025-05-14 at 19 18 53](https://github.com/user-attachments/assets/8cd3232c-53cd-4e10-bc82-277d30fc717c)

Full output reproduced below

----


# Point Queries

A UI may:
1. Need to run many point queries at once (e.g. for a list of items)
2. Have many point queries open at once (e.g. same list of items)

We need to understand the limits of both cases.
- How many point queries can we hydrate at once?
- How many writes per second can we do with many point queries open at once?

## Hydrate 100 point queries

How long to hydrate 100 point queries?
100 seems like a reasonable number. The intuition here is that this
could be a few pages of a list or table view.
1,000 point queries could be possible if each cell in a table grabbed its own data.

  - Total time: `37.30`p99, `36.02`avg
  - Frame budget: 16ms
  - FPS: `0.43`p99, `0.44`avg

### Findings

`2.33`  frames to hydrate 100 point queries which is
~`0.43` fps.

Is this reasonable? Lets check against point lookups from a map.


- Point lookup from map: `0.0035844321`ms.
- Frame budget: 16ms
- Total frames: `0.000224027`
- FPS: `4,463.75`

Point lookup perf is entirely **unreasonable**.
- We can only hydrate 100 points queries at `0.43`fps
- We can do 100 point lookups from a map at `4,463.75`fps


## Maintain 100 Point Queries
How many writes per second can we do with 100 point queries open at once?



  - Maintain queries over 1 write: `0.02`ms p99 `0.02`ms avg.
  - Frame budget: 16ms
  - Writes per frame: `726.1304931928`p99, `765.6366993306`avg

# Range Queries

Single table range queries may or may not be common, depending on how the user structures their app.
A user may do everything as range queries if queries are colocated with components thus pushing the
"join" into the UI.

10 range queries seems like a reasonable number for an app to have open at once.
- A list of tracks
- A list of albums
- A list of artists
- A list of playlists
- A list of liked things
- A list of genres
etc.

The limit on each range query would vary from 10 -> 100.
100 for a main list view, ~10 for all the auxiliary lists.

## Hydrate 10 Range Queries



  - Total time: `1.36`p99, `0.16`avg
  - Frame budget: 16ms
  - FPS: `11.74`p99, `100.10`avg

**Interestingly range query hydration is faster than point query hydration.**
Even if we scaled this up to 100 range queries it would still be faster?

Alternatives to test:
1. Range queries will not always start at the beginning of the table.
2. Range queries will not always be sorted by primary key

## Hydrate 10 Range Queries, Order By Alternate Fields


  - Total time: `1.38`p99, `0.17`avg
  - Frame budget: 16ms
  - FPS: `11.60`p99, `94.40`avg

Surprisingly this is not much slower than order by primary key. That is likely
due to the cost of index creation being paid once up front.

Starting in the middle of a table could still be expensive. The feeling is that it must be
because point queries are so slow.

## Hydrate 10 Range Queries, Start in the Middle of the Table



  - Total time: `1.99`p99, `0.20`avg
  - Frame budget: 16ms
  - FPS: `8.04`p99, `78.55`avg

Something strange. This is still much faster than point queries.
## Scale to 100 Range Queries



  - Total time: `19.92`p99, `7.90`avg
  - Frame budget: 16ms
  - FPS: `0.80`p99, `2.02`avg

Weird. Point query is on average 4.558848358096056 times slower than the range query.

Maybe point query is doing something wrong with limits?

## 100 Point Queries, Limit 1

  - Total time: `3.24`p99, `0.76`avg
  - Frame budget: 16ms
  - FPS: `4.94`p99, `21.19`avg

**TODO: Oops, we're not limiting to 1 by default when we do a point query.**

Now to check maintenance of range queries during writes.

## Maintain 10 Range Queries, Order By Alternate Fields

  - Maintain queries over 1 write: `0.01`ms p99 `0.00`ms avg.
  - Frame budget: 16ms
  - Writes per frame: `2,030.3665205647`p99, `3,464.6928197585`avg

Well this is surprisingly faster than maintaining point queries (w/o limit)!!!
- p99 `2.80`x faster than point queries
- avg `4.53`x faster than point queries

Some other bugs still exist in point queries?